Refactor NodeIterator #29
Merge UserStorageIterator with RecursiveNodeIterator, don't use filter iterator
